+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
   - **Drop Ruby 2.7 support** because there is no `URI:WS` and `URI:WSS` used for `defang_uri` and anyway [official support has ended too](https://www.ruby-lang.org/en/downloads/branches/)
 - Changes:
   - Renamed `from_hexip` to `from_hexipv4` and created an aliases `from_hexip` for `from_hexipv4`. Same for `to_hexipv4` and bang methods.
+- New methods: `from_hexipv6` and `from_hexipv6!`
 
 ## [2.3.0]